% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionOthers.R
\name{add_has_ms2_to_SpecAnnotationClass}
\alias{add_has_ms2_to_SpecAnnotationClass}
\title{add_has_ms2_to_SpecAnnotationClass}
\usage{
add_has_ms2_to_SpecAnnotationClass(obj_class, ms2)
}
\arguments{
\item{obj_class}{a pecAnnotationClass object}
}
\description{
add_has_ms2_to_SpecAnnotationClass
}
\author{
Zhiwei Zhou
}
