% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionOthers.R
\name{add_ms2_purifty_to_SpecAnnotationClass}
\alias{add_ms2_purifty_to_SpecAnnotationClass}
\title{add_ms2_purifty_to_SpecAnnotationClass}
\usage{
add_ms2_purifty_to_SpecAnnotationClass(obj_class, ms2_data_combined)
}
\arguments{
\item{obj_class}{a pecAnnotationClass object}
}
\description{
add_ms2_purifty_to_SpecAnnotationClass
}
\author{
Zhiwei Zhou
}
