% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionDataPrep.R
\name{check_parameters}
\alias{check_parameters}
\title{check_parameters}
\usage{
check_parameters(
  peak_table_unlabel,
  peak_table_label,
  sample_info,
  path = ".",
  raw_data_folder = NULL,
  control_group = c("WT"),
  case_group = c("hyuA")
)
}
\arguments{
\item{peak_table_unlabel}{peak table for unlabeled samples}

\item{peak_table_label}{peak table for labeled samples}

\item{sample_info}{sample information table}

\item{path}{working directory. Default: "."}

\item{control_group}{control group name. Default: "WT"}

\item{case_group}{case group name. Default: "hyuA"}

\item{tracer_condition}{tracer condition, either '12C' or '13C'. Default: '12C'}
}
\description{
check_parameters
}
\examples{
\dontrun{
check_parameters(peak_table_unlabel = 'peak_table_C12.csv',
peak_table_label = 'peak_table_C13.csv',
sample_info = 'sample_info.xlsx',
path = '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data_msdial/',
control_group = c("WT"),
case_group = c('hyuA'))
}
}
\author{
Zhiwei Zhou
}
