% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionOthers.R
\name{combine_ms1_ms2}
\alias{combine_ms1_ms2}
\title{combine_ms1_ms2}
\usage{
combine_ms1_ms2(
  ms1_data,
  ms2_data,
  mz_tol_combine_ms1_ms2 = 15,
  rt_tol_combine_ms1_ms2 = 15,
  ms2_type = c("mgf", "msp", "mzXML", "mzML")
)
}
\arguments{
\item{ms1_data}{The name of ms1 peak table. Column 1 is "name", Column 2 is "mz" and column 3 is "rt".}

\item{ms2_data}{The vector of names of ms2 files. MS2 file must be mzXML.}

\item{mz_tol_combine_ms1_ms2}{mz tol for ms1 and ms2 data matching.}

\item{rt_tol_combine_ms1_ms2}{RT tol for ms1 and ms2 data matching.}

\item{ms2_type}{The type of MS2 file, default is mzXML.}
}
\value{
Return ms1 and ms2 data.
}
\description{
combine_ms1_ms2
}
\author{
Zhiwei Zhou
\email{zhouzw@stanford.edu}
}
