% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionDiffAna.R
\name{diff_analysis}
\alias{diff_analysis}
\title{diff_analysis}
\usage{
diff_analysis(
  peak_table,
  sample_info,
  control_group = "WT",
  case_group = "hyuA",
  tracer_condition = c("12C", "13C"),
  p_value_cutoff = 0.05,
  p_adjust = TRUE,
  fold_change_cutoff = 20,
  abundance_cutoff = 10^4
)
}
\arguments{
\item{peak_table}{peak table}

\item{sample_info}{sample information table}

\item{control_group}{control_group label. Default: "WT"}

\item{case_group}{case_group label. Default: "hyuA"}

\item{tracer_condition}{tracer condition, either '12C' or '13C'. Default: c('12C', '13C')}

\item{p_value_cutoff}{Default: 0.05}

\item{p_adjust}{Default: TRUE}

\item{fold_change_cutoff}{Default: 20}
}
\description{
diff_analysis
}
\examples{
\dontrun{
diff_analysis(peak_table = '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data/hyuA/C12/converted_peak_table.csv',
              sample_info = '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data/hyuA/sample_info.xlsx',
              control_group = "WT",
              case_group = 'hyuA',
              tracer_condition = '12C',
              p_value_cutoff = 0.05,
              p_adjust = TRUE,
              fold_change_cutoff = 20)
}
}
\author{
Zhiwei Zhou
}
