% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{find_pair_feature}
\alias{find_pair_feature}
\title{find_pair_feature}
\usage{
find_pair_feature(
  target_mz,
  target_rt,
  polarity,
  source_feature_id,
  labeled_feature_table,
  mz_tol,
  rt_tol
)
}
\arguments{
\item{target_mz}{target mz value}

\item{target_rt}{target retention time}

\item{labeled_feature_table}{labeled feature table}

\item{mz_tol}{mz tolerance in ppm}

\item{rt_tol}{retention time tolerance in minutes}
}
\value{
A data frame containing the matched feature information
}
\description{
find_pair_feature
}
\author{
Zhiwei Zhou
}
