% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionPeakGrouping.R
\name{generatePeakGroup}
\alias{generatePeakGroup}
\title{generatePeakGroup}
\usage{
generatePeakGroup(
  peak_table,
  base_peak_name,
  raw_data_files,
  base_peak_adduct = c("[M+H]+", "[M-H]-"),
  tol_mz = 20,
  tol_rt = 0.05,
  cutoff_ssc = 0.8,
  cutoff_ppc = 0.8,
  is_calculate_ppc = TRUE,
  path = "."
)
}
\arguments{
\item{peak_table}{peak table (row--peaks; column--samples). First 3 column should be named as "name", "mz", "rt"}

\item{base_peak_name}{name of base peak}

\item{base_peak_adduct}{adduct format of base peak}

\item{tol_rt}{rt tolerance for peak grouping; Default: 3 (second)}

\item{cutoff_ssc}{sample-sample correlation; Default: 0}

\item{base_peak_seed}{seed of base peak}
}
\description{
generatePeakGroup
}
\author{
Zhiwei Zhou
}
