% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionDataPrep.R
\name{modify_msdial_table}
\alias{modify_msdial_table}
\title{modify_msdial_table}
\usage{
modify_msdial_table(
  table_msdial,
  table_sample_info,
  path = ".",
  file_output = "converted_peak_table.csv"
)
}
\arguments{
\item{table_msdial}{peak table from xcms}

\item{table_sample_info}{sample information table}

\item{path}{working directory. Default: "."}

\item{file_output}{'converted_peak_table.csv'}
}
\description{
modify_msdial_table
}
\examples{
\dontrun{
table_msdial <- '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data_msdial/hyuA_UA_48h_area.txt'
table_sample_info <- '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data_msdial/sample_info.xlsx'
modify_msdial_table(table_msdial = table_msdial,
                     table_sample_info = table_sample_info,
                     path = '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data_msdial',
                     file_output = 'converted_peak_table.csv')
}
}
\author{
Zhiwei Zhou
}
