% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionDataPrep.R
\name{modify_mzmine_table}
\alias{modify_mzmine_table}
\title{modify_mzmine_table}
\usage{
modify_mzmine_table(
  table_mzmine,
  table_sample_info,
  path = ".",
  file_output = "converted_peak_table.csv"
)
}
\arguments{
\item{table_mzmine}{peak table from xcms}

\item{table_sample_info}{sample information table}

\item{path}{working directory. Default: "."}

\item{file_output}{'converted_peak_table.csv'}
}
\description{
modify_mzmine_table
}
\examples{
\dontrun{
table_mzmine <- '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data_mzmine/mzmine_12C.csv'
table_sample_info <- '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data_mzmine/sample_info.xlsx'
modify_mzmine_table(table_mzmine = table_mzmine,
                     table_sample_info = table_sample_info,
                     path = '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data_mzmine',
                     file_output = 'converted_peak_table.csv')
}
}
\author{
Zhiwei Zhou
}
