% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionDataPrep.R
\name{modify_tidymass_table}
\alias{modify_tidymass_table}
\title{modify_tidymass_table}
\usage{
modify_tidymass_table(
  table_tidymass,
  table_camera,
  table_sample_info,
  path = ".",
  impute_NA = TRUE,
  impute_NA_value = 100,
  polarity = c("positive", "negative"),
  file_output = "converted_peak_table.csv"
)
}
\arguments{
\item{table_tidymass}{peak table from tidymass}

\item{table_camera}{isotopes and adducts annotated by the CAMERA}

\item{table_sample_info}{sample information table}

\item{path}{working directory. Default: "."}

\item{file_output}{'converted_peak_table.csv'}
}
\description{
modify_tidymass_table
}
\examples{
\dontrun{
table_tidymass <- '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data_tidymass/00_raw_data_processing_12C/Result/Peak_table_for_cleaning.csv'
table_camera <- '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data_tidymass/00_raw_data_processing_12C/00_raw_data_processing/adduct_result_camera.xlsx'
table_sample_info <- '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data_tidymass/sample_info.xlsx'
modify_tidymass_table(table_tidymass = table_tidymass,
                      table_camera = table_camera,
                      table_sample_info = table_sample_info,
                  path = '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data/hyuA/C12')
}
}
\author{
Zhiwei Zhou
}
