% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionDataPrep.R
\name{modify_xcms_table}
\alias{modify_xcms_table}
\title{modify_xcms_table}
\usage{
modify_xcms_table(
  table_xcms,
  table_camera,
  table_sample_info,
  path = ".",
  file_output = "converted_peak_table.csv"
)
}
\arguments{
\item{table_xcms}{peak table from xcms}

\item{table_camera}{isotopes and adducts annotated by the CAMERA}

\item{table_sample_info}{sample information table}

\item{path}{working directory. Default: "."}

\item{file_output}{'converted_peak_table.csv'}
}
\description{
modify_xcms_table
}
\examples{
\dontrun{
table_xcms <- '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data/hyuA/C12/00_raw_data_processing/Peak-table.csv'
table_camera <- '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data/hyuA/C12/00_raw_data_processing/adduct_result_camera.xlsx'
table_sample_info <- '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data/hyuA/sample_info.xlsx'
modify_xcms_table(table_xcms = table_xcms,
                  table_camera = table_camera,
                  table_sample_info = table_sample_info,
                  path = '~/Project/00_Uric_Acid_project/Data/20250606_isopairfind_test/Demo_data/hyuA/C12')
}
}
\author{
Zhiwei Zhou
}
