% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionOthers.R
\name{read_ms2}
\alias{read_ms2}
\title{read_ms2}
\usage{
read_ms2(ms2_file, ms2_type = c("mgf", "cef", "msp", "mzXML", "mzML"), ...)
}
\arguments{
\item{ms2_file}{MS/MS file names}

\item{ms2_type}{'mgf', 'cef', 'msp', 'mzXML', 'mzML'}
}
\description{
read_ms2
}
\author{
Zhiwei Zhou
}
