% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionOthers.R
\name{readMzML2}
\alias{readMzML2}
\title{readMzML2}
\usage{
readMzML2(
  file,
  isolation_window = c("narrow", "medium", "wide"),
  is_purification = FALSE
)
}
\arguments{
\item{file}{mzml or mzxml file. ms2 files}

\item{isolation_window}{'narrow', 'medium', 'wide'}

\item{is_purification}{whether do the ms2 purification}
}
\description{
readMzML2
}
\examples{
\dontrun{
file <- '~/Project/00_IBD_project/Data/20230428_Zhiwei_StanfordPool_MS2/mzxml_haoqing_data/Haoqing_MS2_data/MS2_lib_mzXML/Stanford_HILIC_pos/SU-A01_1.mzXML'
file <- '~/Project/00_IBD_project/Data/20230428_Zhiwei_StanfordPool_MS2/mzml_hilic_pos/StanfordPool1-2.mzML'

test <- retrieve_raw_ms2(file = 'SU-A01_1.mzXML',
                         isolation_window = 'narrow',
                         dir_path = '~/Project/00_IBD_project/Data/20230428_Zhiwei_StanfordPool_MS2/mzxml_haoqing_data/Haoqing_MS2_data/MS2_lib_mzXML/Stanford_HILIC_pos')
}
}
\author{
Zhiwei Zhou
}
