% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionPeakGrouping.R
\name{recognize_rela_peak}
\alias{recognize_rela_peak}
\title{recognize_rela_peak}
\usage{
recognize_rela_peak(
  peak_table,
  peak_table_signif,
  path_dir = ".",
  polarity = c("positive", "negative"),
  raw_data_folder = ".",
  tol_mz = 20,
  tol_rt = 0.05,
  cutoff_ssc = 0.8,
  cutoff_ppc = 0.8
)
}
\arguments{
\item{polarity}{"positive", "negative"}
}
\description{
recognize related peaks for enriched peaks, e.g. adducts, neutral losses, in-source fragments
}
\author{
Zhiwei Zhou
}
