% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionPlot.R
\name{ZZW_annotate_text2}
\alias{ZZW_annotate_text2}
\title{ZZW_annotate_text2}
\usage{
ZZW_annotate_text2(
  label,
  x,
  y,
  facets = NULL,
  hjust = 0,
  vjust = 0,
  color = "black",
  alpha = NA,
  family = thm$text$family,
  size = thm$text$size,
  fontface = 1,
  lineheight = 1,
  box_just = ifelse(c(x, y) < 0.5, 0, 1),
  margin = grid::unit(size/2, "pt"),
  thm = ggplot2::theme_get()
)
}
\description{
ZZW_annotate_text2
}
\examples{
library(ggplot2)
ggplot(diamonds, aes(carat, price, color = cut)) +
  geom_point() +
  annotate(geom = 'text', x = 4, y = 5000,
           label = "a\nb\nc")
qplot(1:10,1:10) + ZZW_annotate_text2('some long text\nx = 1', x=1, y=0, hjust=1)
}
\author{
Zhiwei Zhou
}
